+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/trackselection/DefaultTrackSelector.java
@@ -44,6 +44,7 @@ import androidx.annotation.Nullable;
 import androidx.media3.common.AudioAttributes;
 import androidx.media3.common.C;
 import androidx.media3.common.C.RoleFlags;
+import androidx.media3.common.ColorInfo;
 import androidx.media3.common.Format;
 import androidx.media3.common.MimeTypes;
 import androidx.media3.common.Timeline;
@@ -63,6 +64,7 @@ import androidx.media3.exoplayer.RendererCapabilities.AdaptiveSupport;
 import androidx.media3.exoplayer.RendererCapabilities.Capabilities;
 import androidx.media3.exoplayer.RendererConfiguration;
 import androidx.media3.exoplayer.audio.AudioSink;
+import androidx.media3.exoplayer.mediacodec.MediaCodecUtil;
 import androidx.media3.exoplayer.source.MediaSource.MediaPeriodId;
 import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.util.SpatializerWrapper;
@@ -3833,9 +3835,12 @@ public class DefaultTrackSelector extends MappingTrackSelector
     private final int selectedAudioLanguageScore;
     private final boolean allowMixedMimeTypes;
     private final @SelectionEligibility int selectionEligibility;
+    private final boolean usesPrimaryOrFallbackDecoder;
     private final boolean usesPrimaryDecoder;
     private final boolean usesHardwareAcceleration;
     private final int codecPreferenceScore;
+    private final boolean isHdr;
+    @Nullable private final String resolvedMimeType;
 
     public VideoTrackInfo(
         int rendererIndex,
@@ -3902,23 +3907,37 @@ public class DefaultTrackSelector extends MappingTrackSelector
           normalizeUndeterminedLanguageToNull(selectedAudioLanguage) == null;
       selectedAudioLanguageScore =
           getFormatLanguageScore(format, selectedAudioLanguage, selectedAudioLanguageUndetermined);
+
+      String resolvedMimeType = format.sampleMimeType;
+      @RendererCapabilities.DecoderSupport
+      int decoderSupport = RendererCapabilities.getDecoderSupport(formatSupport);
+      if (decoderSupport == RendererCapabilities.DECODER_SUPPORT_FALLBACK_MIMETYPE) {
+        String fallbackMimeType = MediaCodecUtil.getAlternativeCodecMimeType(format);
+        if (fallbackMimeType != null) {
+          resolvedMimeType = fallbackMimeType;
+        }
+      }
+
       int bestMimeTypeMatchIndex = Integer.MAX_VALUE;
       for (int i = 0; i < parameters.preferredVideoMimeTypes.size(); i++) {
-        if (format.sampleMimeType != null
-            && format.sampleMimeType.equals(parameters.preferredVideoMimeTypes.get(i))) {
+        if (resolvedMimeType != null
+            && resolvedMimeType.equals(parameters.preferredVideoMimeTypes.get(i))) {
           bestMimeTypeMatchIndex = i;
           break;
         }
       }
       preferredMimeTypeMatchIndex = bestMimeTypeMatchIndex;
       preferredLabelMatchIndex = getBestLabelMatchIndex(format, parameters.preferredVideoLabels);
-      usesPrimaryDecoder =
-          RendererCapabilities.getDecoderSupport(formatSupport)
-              == RendererCapabilities.DECODER_SUPPORT_PRIMARY;
+      usesPrimaryOrFallbackDecoder =
+          decoderSupport == RendererCapabilities.DECODER_SUPPORT_PRIMARY
+              || decoderSupport == RendererCapabilities.DECODER_SUPPORT_FALLBACK_MIMETYPE;
+      usesPrimaryDecoder = decoderSupport == RendererCapabilities.DECODER_SUPPORT_PRIMARY;
       usesHardwareAcceleration =
           RendererCapabilities.getHardwareAccelerationSupport(formatSupport)
               == RendererCapabilities.HARDWARE_ACCELERATION_SUPPORTED;
-      codecPreferenceScore = getVideoCodecPreferenceScore(format.sampleMimeType);
+      this.resolvedMimeType = resolvedMimeType;
+      codecPreferenceScore = getVideoCodecPreferenceScore(resolvedMimeType);
+      isHdr = usesPrimaryDecoder && ColorInfo.isTransferHdr(format.colorInfo);
       selectionEligibility = evaluateSelectionEligibility(formatSupport, requiredAdaptiveSupport);
     }
 
@@ -3930,9 +3949,9 @@ public class DefaultTrackSelector extends MappingTrackSelector
     @Override
     public boolean isCompatibleForAdaptationWith(VideoTrackInfo otherTrack) {
       return (allowMixedMimeTypes
-              || Objects.equals(format.sampleMimeType, otherTrack.format.sampleMimeType))
+              || Objects.equals(this.resolvedMimeType, otherTrack.resolvedMimeType))
           && (parameters.allowVideoMixedDecoderSupportAdaptiveness
-              || (this.usesPrimaryDecoder == otherTrack.usesPrimaryDecoder
+              || (this.usesPrimaryOrFallbackDecoder == otherTrack.usesPrimaryOrFallbackDecoder
                   && this.usesHardwareAcceleration == otherTrack.usesHardwareAcceleration));
     }
 
@@ -3988,12 +4007,9 @@ public class DefaultTrackSelector extends MappingTrackSelector
                   info2.preferredMimeTypeMatchIndex,
                   Ordering.natural().reverse())
               // 5. Compare match with renderer capability preferences.
-              .compareFalseFirst(info1.usesPrimaryDecoder, info2.usesPrimaryDecoder)
+              .compareFalseFirst(
+                  info1.usesPrimaryOrFallbackDecoder, info2.usesPrimaryOrFallbackDecoder)
               .compareFalseFirst(info1.usesHardwareAcceleration, info2.usesHardwareAcceleration);
-
-      if (info1.usesPrimaryDecoder && info1.usesHardwareAcceleration) {
-        chain = chain.compare(info1.codecPreferenceScore, info2.codecPreferenceScore);
-      }
       return chain.result();
     }
 
@@ -4015,8 +4031,17 @@ public class DefaultTrackSelector extends MappingTrackSelector
         comparisonChain =
             comparisonChain.compare(info1.bitrate, info2.bitrate, FORMAT_VALUE_ORDERING.reverse());
       }
-      return comparisonChain
-          .compare(info1.pixelCount, info2.pixelCount, qualityOrdering)
+      ComparisonChain chain =
+          comparisonChain
+              .compareFalseFirst(info1.isHdr, info2.isHdr)
+              .compare(info1.pixelCount, info2.pixelCount, qualityOrdering);
+
+      if (info1.usesPrimaryOrFallbackDecoder && info1.usesHardwareAcceleration) {
+        chain = chain.compare(info1.codecPreferenceScore, info2.codecPreferenceScore);
+      }
+
+      return chain
+          .compareFalseFirst(info1.usesPrimaryDecoder, info2.usesPrimaryDecoder)
           .compare(info1.bitrate, info2.bitrate, qualityOrdering)
           .result();
     }